Fix trace lookup and export validation
