+++ apps/api/src/schemas/export.ts
@@ -1,14 +1,16 @@
 import * as v from 'valibot';
 
-import { tsParam } from '../utils/valibot.js';
+import { toNum } from '../utils/valibot.js';
 
 export const ExportFormatSchema = v.picklist(['json', 'csv', 'text']);
+
+const ExportTimestamp = v.pipe(toNum, v.minValue(0));
 
 export const ExportLogsQuery = v.pipe(
 	v.object({
 		q: v.optional(v.string()),
-		startTs: tsParam,
-		endTs: tsParam,
+		startTs: ExportTimestamp,
+		endTs: ExportTimestamp,
 		format: ExportFormatSchema,
 		dryRun: v.optional(
 			v.pipe(
